+++ source/components/dispatcher/dswstate.c
@@ -189,7 +189,7 @@ AcpiDsResultPop (
     /* Return object of the top element and clean that top element result stack */
 
     WalkState->ResultCount--;
-    Index = WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
+    Index = (UINT32) WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
 
     *Object = State->Results.ObjDesc [Index];
     if (!*Object)
@@ -284,7 +284,7 @@ AcpiDsResultPush (
 
     /* Assign the address of object to the top free element of result stack */
 
-    Index = WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
+    Index = (UINT32) WalkState->ResultCount % ACPI_RESULTS_FRAME_OBJ_NUM;
     State->Results.ObjDesc [Index] = Object;
     WalkState->ResultCount++;
